fix: ui overlap
